+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- add SwiftUI wrapper `StepsView`.
+
 ### Fixed
 - calculate `intrinsicContentSize`.